+++ hs-bindgen/src-internal/HsBindgen/Config.hs
@@ -14,7 +14,6 @@ module HsBindgen.Config (
 where
 
 import HsBindgen.Backend.Category
-import HsBindgen.Backend.Hs.Haddock.Config
 import HsBindgen.BindingSpec
 import HsBindgen.Config.ClangArgs
 import HsBindgen.Config.Internal
@@ -46,8 +45,6 @@ data Config_ path = Config {
   , fieldNamingStrategy :: FieldNamingStrategy
   , emptyMacros         :: EmptyMacros
 
-    -- * Backend
-  , haddockPathStyle :: PathStyle
   }
   deriving stock (Eq, Show, Generic)
   deriving stock (Functor, Foldable, Traversable)
@@ -75,7 +72,6 @@ toBindgenConfig config uniqueId baseModuleName choice =
         }
     , backend = BackendConfig {
           uniqueId       = uniqueId
-        , haddock        = def & #pathStyle .~ config.haddockPathStyle
         , categoryChoice = choice
        }
     }